Move Affine3DUtils to backend package
