+++ src/main/java/sc/fiji/ome/zarr/pyramid/backend/Affine3DUtils.java
@@ -26,7 +26,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package sc.fiji.ome.zarr.pyramid;
+package sc.fiji.ome.zarr.pyramid.backend;
 
 import net.imglib2.realtransform.AffineTransform3D;
